backend/requirements.txt
# 🚀 Yojna AI Backend Requirements (Fixed)

fastapi==0.115.5
uvicorn[standard]==0.32.1

# MongoDB (Fixed compatibility)
# Do NOT `pip install bson` — the standalone "bson" PyPI package conflicts with pymongo.

motor==3.6.0
pymongo==4.9.2   # 👈 IMPORTANT FIX

# Gemini AI

google-generativeai==0.8.3

# Env

pydantic-settings==2.6.1
python-dotenv==1.0.1

# Validation

pydantic==2.10.3

# Utils


httpx==0.27.2

# Dev

pytest==8.3.3
pytest-asyncio==0.24.0
